Infinite load, can't find lance bug fix
Under weird circumstances it can ask for a lance at too high difficulty, over difficulty 10, and for unknown reasons it only wants to grab lances within 10 difficulty matching the lance include/excludes.  If it doesn't find a lance within 10 difficulty of the lance it will totally crap out further contract generation and cause an infinite load.  Allowing over 10 difference in difficulty in it's search in this method should fix the issue.  This isn't a Mission Control specific fix as other mod set ups could cause this but Mission Control with certain settings could induce this issue in of itself so worth fixing here.